styles in hoprhunt fixed
